Repair broken passphrase retry
The exit code handling of the gpg process to detect wrong passphrases
was a bit adventurous. It checked for the raw status value from waitpid
instead of the WEXITSTATUS() macro.
The code that handles the waitpid also does not check for any errors.
This is now fixed by evaluating the waitpid return value and the use of
the WEXITSTATUS() macro.

Signed-off-by: Jan Losinski <losinski@wh2.tu-dresden.de>